Replace signed bitwise operations in march_hardware/imotioncube with unsigned operations

+++ ros1/src/hardware/interface/march_hardware/src/motor_controller/imotioncube/imotioncube.cpp
@@ -468,7 +468,7 @@ bool IMotionCube::verifySetup(SdoSlaveInterface& sdo)
         = this->computeSWCheckSum(start_address, end_address);
     // set parameters to compute checksum on the imc
     const int checksum_setup = sdo.write<uint32_t>(
-        /*index=*/0x2069, /*sub=*/0, (end_address << 16) | start_address);
+        /*index=*/0x2069, /*sub=*/0, (end_address << 16U) | start_address);
 
     uint16_t imc_value;
     int value_size = sizeof(imc_value);
@@ -488,8 +488,8 @@ bool IMotionCube::verifySetup(SdoSlaveInterface& sdo)
 
 void IMotionCube::downloadSetupToDrive(SdoSlaveInterface& sdo)
 {
-    int result = 0;
-    int final_result = 0;
+    uint32_t result = 0;
+    uint32_t final_result = 0;
 
     size_t pos = 0;
     size_t old_pos = 0;
@@ -500,8 +500,8 @@ void IMotionCube::downloadSetupToDrive(SdoSlaveInterface& sdo)
             const uint16_t mem_location
                 = std::stoi(token, /*__idx=*/nullptr, /*__base=*/16);
             const uint16_t mem_setup = 9; // send 16-bits and auto increment
-            result = sdo.write<uint32_t>(/*index=*/0x2064, /*sub=*/0,
-                (mem_location << 16)
+            result = (uint32_t)sdo.write<uint32_t>(/*index=*/0x2064, /*sub=*/0,
+                (mem_location << 16U)
                     | mem_setup); // write the write-configuration
             final_result |= result;
         } else {
@@ -515,14 +515,16 @@ void IMotionCube::downloadSetupToDrive(SdoSlaveInterface& sdo)
 
                 uint32_t data = 0;
                 if (pos != old_pos) {
-                    data = (std::stoi(
+                    data = ((uint32_t)std::stoi(
                                 next_token, /*__idx=*/nullptr, /*__base=*/16)
-                               << 16)
-                        | std::stoi(token, /*__idx=*/nullptr, /*__base=*/16);
+                               << 16U)
+                        | (uint32_t)std::stoi(
+                            token, /*__idx=*/nullptr, /*__base=*/16);
                 } else {
-                    data = std::stoi(token, /*__idx=*/nullptr, /*__base=*/16);
+                    data = (uint32_t)std::stoi(
+                        token, /*__idx=*/nullptr, /*__base=*/16);
                 }
-                result = sdo.write<uint32_t>(
+                result = (uint32_t)sdo.write<uint32_t>(
                     /*index=*/0x2065, /*sub=*/0,
                     data); // write the write-configuration
             }
